Creating abstract class PersonBase for both Member and Staff
