+++ src/main/java/com/ros/lms/domain/entities/PersonBase.java
@@ -1,0 +1,97 @@
+package com.ros.lms.domain.entities;
+
+import com.ros.lms.domain.enums.Sex;
+import jakarta.persistence.*;
+
+@MappedSuperclass
+public abstract class PersonBase {
+
+    @Column(name = "GOVERNMENT_ID")
+    private String governmentID;
+
+    @Column(name = "FIRST_NAME")
+    private String firstName;
+
+    @Column(name = "MIDDLE_NAME")
+    private String middleName;
+
+    @Column(name = "LAST_NAME")
+    private String lastName;
+
+    @Column(name = "PHONE")
+    private String phone;
+
+    @Column(name = "SEX")
+    private Sex sex;
+
+    @Column(name = "EMAIL")
+    private String email;
+
+    @Column(name = "USERNAME")
+    private String username;
+
+
+    public String getGovernmentID() {
+        return governmentID;
+    }
+
+    public void setGovernmentID(String governmentID) {
+        this.governmentID = governmentID;
+    }
+
+    public String getFirstName() {
+        return firstName;
+    }
+
+    public void setFirstName(String firstName) {
+        this.firstName = firstName;
+    }
+
+    public String getMiddleName() {
+        return middleName;
+    }
+
+    public void setMiddleName(String middleName) {
+        this.middleName = middleName;
+    }
+
+    public String getLastName() {
+        return lastName;
+    }
+
+    public void setLastName(String lastName) {
+        this.lastName = lastName;
+    }
+
+    public String getPhone() {
+        return phone;
+    }
+
+    public void setPhone(String phone) {
+        this.phone = phone;
+    }
+
+    public Sex getSex() {
+        return sex;
+    }
+
+    public void setSex(Sex sex) {
+        this.sex = sex;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
+    public String getUsername() {
+        return username;
+    }
+
+    public void setUsername(String username) {
+        this.username = username;
+    }
+}
