+++ src/main/java/com/ros/lms/domain/entities/Staff.java
@@ -5,35 +5,11 @@ import jakarta.persistence.*;
 
 @Entity
 @Table(name = "STAFFS")
-public class Staff {
+public class Staff extends PersonBase{
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "STAFF_ID")
     private long id;
-
-    @Column(name = "GOVERNMENT_ID")
-    private String governmentID;
-
-    @Column(name = "FIRST_NAME")
-    private String firstName;
-
-    @Column(name = "MIDDLE_NAME")
-    private String middleName;
-
-    @Column(name = "LAST_NAME")
-    private String lastName;
-
-    @Column(name = "PHONE")
-    private String phone;
-
-    @Column(name = "SEX")
-    private Sex sex;
-
-    @Column(name = "EMAIL")
-    private String email;
-
-    @Column(name = "USERNAME")
-    private String username;
 
     public Staff() {}
 
@@ -101,14 +77,14 @@ public class Staff {
 
     private Staff(Builder builder) {
         id = builder.id;
-        governmentID = builder.governmentID;
-        firstName = builder.firstName;
-        middleName = builder.middleName;
-        lastName = builder.lastName;
-        phone = builder.phone;
-        sex = builder.sex;
-        email = builder.email;
-        username = builder.username;
+        super.setGovernmentID(builder.governmentID);
+        super.setFirstName(builder.firstName);
+        super.setMiddleName(builder.middleName);
+        super.setLastName(builder.lastName);
+        super.setPhone(builder.phone);
+        super.setSex(builder.sex);
+        super.setEmail(builder.email);
+        super.setUsername(builder.username);
     }
 
     public long getId() {
@@ -119,82 +95,18 @@ public class Staff {
         this.id = id;
     }
 
-    public String getGovernmentID() {
-        return governmentID;
-    }
-
-    public void setGovernmentID(String governmentID) {
-        this.governmentID = governmentID;
-    }
-
-    public String getFirstName() {
-        return firstName;
-    }
-
-    public void setFirstName(String firstName) {
-        this.firstName = firstName;
-    }
-
-    public String getLastName() {
-        return lastName;
-    }
-
-    public void setLastName(String lastName) {
-        this.lastName = lastName;
-    }
-
-    public String getMiddleName() {
-        return middleName;
-    }
-
-    public void setMiddleName(String middleName) {
-        this.middleName = middleName;
-    }
-
-    public String getPhone() {
-        return phone;
-    }
-
-    public void setPhone(String phone) {
-        this.phone = phone;
-    }
-
-    public Sex getSex() {
-        return sex;
-    }
-
-    public void setSex(Sex sex) {
-        this.sex = sex;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
     @Override
     public String toString() {
         return "Staff{" +
                 "id=" + id +
-                ", governmentID='" + governmentID + '\'' +
-                ", firstName='" + firstName + '\'' +
-                ", middleName='" + middleName + '\'' +
-                ", lastName='" + lastName + '\'' +
-                ", phone='" + phone + '\'' +
-                ", sex=" + sex +
-                ", email='" + email + '\'' +
-                ", username='" + username + '\'' +
+                ", governmentID='" + super.getGovernmentID() + '\'' +
+                ", firstName='" + super.getFirstName() + '\'' +
+                ", middleName='" + super.getMiddleName() + '\'' +
+                ", lastName='" + super.getLastName() + '\'' +
+                ", phone='" + super.getPhone() + '\'' +
+                ", sex=" + (super.getSex() != null ? super.getSex().getCode() : "null") +
+                ", email='" + super.getEmail() + '\'' +
+                ", username='" + super.getUsername() + '\'' +
                 '}';
     }
 }
